+++ library/oci_autonomous_database.py
@@ -75,6 +75,11 @@ options:
         description: The destination file path with file name when downloading wallet. The file must have 'zip' extension.
                      I(wallet_file) is required if I(state='generate_wallet').
         required: false
+    is_free_tier:
+        description: Indicates if this is an Always Free resource. The default value is false. Note that Always Free
+                     Autonomous Databases have 1 CPU and 20GB of memory. For Always Free databases, memory and CPU
+                     cannot be scaled.
+        required: false
     force:
         description: Force overwriting existing wallet file when downloading wallet.
         required: false
@@ -230,6 +235,13 @@ RETURN = """
                 returned: always
                 type: string
                 sample: AVAILABLE
+            is_free_tier:
+                description: Indicates if this is an Always Free resource. The default value is false. Note that Always Free
+                             Autonomous Databases have 1 CPU and 20GB of memory. For Always Free databases, memory and CPU
+                             cannot be scaled.
+                returned: always
+                type: bool
+                sample: false
 
         sample: {
                   "compartment_id":"ocid1.compartment.oc1..xxxxxEXAMPLExxxxx",
@@ -257,7 +269,8 @@ RETURN = """
                   "service_console_url":"https://example1.oraclecloud.com/console/index.html?
                         tenant_name=OCID1.TENANCY.OC1..xxxxxEXAMPLExxxxx
                         &database_name=ANSIBLEAUTODB&service_type=ATP",
-                  "time_created":"2018-09-22T15:06:55.426000+00:00"
+                  "time_created":"2018-09-22T15:06:55.426000+00:00",
+                  "is_free_tier": false
               }
 """
 
@@ -522,6 +535,7 @@ def main():
             ),
             timestamp=dict(type="str", required=False),
             wallet_file=dict(type="str", required=False),
+            is_free_tier=dict(type="bool", required=False),
             force=dict(
                 type="bool", required=False, default=True, aliases=["overwrite"]
             ),