+++ library/oci_budget_alert_rule.py
@@ -4,6 +4,7 @@
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 # Apache License v2.0
 # See LICENSE.TXT for details.
+# GENERATED FILE - DO NOT EDIT - MANUAL CHANGES WILL BE OVERWRITTEN
 
 
 from __future__ import absolute_import, division, print_function
@@ -243,7 +244,7 @@ budget_alert_rule:
 """
 
 from ansible.module_utils.basic import AnsibleModule
-from ansible.module_utils.oracle import oci_common_utils
+from ansible.module_utils.oracle import oci_common_utils, oci_wait_utils
 from ansible.module_utils.oracle.oci_resource_utils import (
     OCIResourceHelperBase,
     get_custom_class,
@@ -268,6 +269,9 @@ class BudgetAlertRuleHelperGen(OCIResourceHelperBase):
 
     def get_module_resource_id(self):
         return self.module.params.get("alert_rule_id")
+
+    def get_get_fn(self):
+        return self.client.get_alert_rule
 
     def get_resource(self):
         return oci_common_utils.call_with_backoff(
@@ -306,10 +310,19 @@ class BudgetAlertRuleHelperGen(OCIResourceHelperBase):
 
     def create_resource(self):
         create_details = self.get_create_model()
-        return oci_common_utils.call_with_backoff(
-            self.client.create_alert_rule,
-            budget_id=self.module.params.get("budget_id"),
-            create_alert_rule_details=create_details,
+        return oci_wait_utils.call_and_wait(
+            call_fn=self.client.create_alert_rule,
+            call_fn_args=(),
+            call_fn_kwargs=dict(
+                budget_id=self.module.params.get("budget_id"),
+                create_alert_rule_details=create_details,
+            ),
+            waiter_type=oci_wait_utils.LIFECYCLE_STATE_WAITER_KEY,
+            operation=oci_common_utils.CREATE_OPERATION_KEY,
+            waiter_client=self.client,
+            resource_helper=self,
+            wait_for_states=self.module.params.get("wait_until")
+            or oci_common_utils.get_resource_active_states(),
         )
 
     def get_update_model_class(self):
@@ -317,18 +330,36 @@ class BudgetAlertRuleHelperGen(OCIResourceHelperBase):
 
     def update_resource(self):
         update_details = self.get_update_model()
-        return oci_common_utils.call_with_backoff(
-            self.client.update_alert_rule,
-            budget_id=self.module.params.get("budget_id"),
-            alert_rule_id=self.module.params.get("alert_rule_id"),
-            update_alert_rule_details=update_details,
+        return oci_wait_utils.call_and_wait(
+            call_fn=self.client.update_alert_rule,
+            call_fn_args=(),
+            call_fn_kwargs=dict(
+                budget_id=self.module.params.get("budget_id"),
+                alert_rule_id=self.module.params.get("alert_rule_id"),
+                update_alert_rule_details=update_details,
+            ),
+            waiter_type=oci_wait_utils.LIFECYCLE_STATE_WAITER_KEY,
+            operation=oci_common_utils.UPDATE_OPERATION_KEY,
+            waiter_client=self.client,
+            resource_helper=self,
+            wait_for_states=self.module.params.get("wait_until")
+            or oci_common_utils.get_resource_active_states(),
         )
 
     def delete_resource(self):
-        return oci_common_utils.call_with_backoff(
-            self.client.delete_alert_rule,
-            budget_id=self.module.params.get("budget_id"),
-            alert_rule_id=self.module.params.get("alert_rule_id"),
+        return oci_wait_utils.call_and_wait(
+            call_fn=self.client.delete_alert_rule,
+            call_fn_args=(),
+            call_fn_kwargs=dict(
+                budget_id=self.module.params.get("budget_id"),
+                alert_rule_id=self.module.params.get("alert_rule_id"),
+            ),
+            waiter_type=oci_wait_utils.NONE_WAITER_KEY,
+            operation=oci_common_utils.DELETE_OPERATION_KEY,
+            waiter_client=self.client,
+            resource_helper=self,
+            wait_for_states=self.module.params.get("wait_until")
+            or oci_common_utils.get_resource_terminated_states(),
         )
 
 
@@ -369,6 +400,7 @@ def main():
         module=module,
         resource_type="budget_alert_rule",
         service_client_class=BudgetClient,
+        namespace="budget",
     )
 
     result = dict(changed=False)