+++ library/oci_network_security_group.py
@@ -4,6 +4,7 @@
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 # Apache License v2.0
 # See LICENSE.TXT for details.
+# GENERATED FILE - DO NOT EDIT - MANUAL CHANGES WILL BE OVERWRITTEN
 
 
 from __future__ import absolute_import, division, print_function
@@ -167,7 +168,7 @@ network_security_group:
 """
 
 from ansible.module_utils.basic import AnsibleModule
-from ansible.module_utils.oracle import oci_common_utils
+from ansible.module_utils.oracle import oci_common_utils, oci_wait_utils
 from ansible.module_utils.oracle.oci_resource_utils import (
     OCIResourceHelperBase,
     get_custom_class,
@@ -192,6 +193,9 @@ class NetworkSecurityGroupHelperGen(OCIResourceHelperBase):
 
     def get_module_resource_id(self):
         return self.module.params.get("network_security_group_id")
+
+    def get_get_fn(self):
+        return self.client.get_network_security_group
 
     def get_resource(self):
         return oci_common_utils.call_with_backoff(
@@ -231,9 +235,16 @@ class NetworkSecurityGroupHelperGen(OCIResourceHelperBase):
 
     def create_resource(self):
         create_details = self.get_create_model()
-        return oci_common_utils.call_with_backoff(
-            self.client.create_network_security_group,
-            create_network_security_group_details=create_details,
+        return oci_wait_utils.call_and_wait(
+            call_fn=self.client.create_network_security_group,
+            call_fn_args=(),
+            call_fn_kwargs=dict(create_network_security_group_details=create_details),
+            waiter_type=oci_wait_utils.LIFECYCLE_STATE_WAITER_KEY,
+            operation=oci_common_utils.CREATE_OPERATION_KEY,
+            waiter_client=self.client,
+            resource_helper=self,
+            wait_for_states=self.module.params.get("wait_until")
+            or oci_common_utils.get_resource_active_states(),
         )
 
     def get_update_model_class(self):
@@ -241,20 +252,38 @@ class NetworkSecurityGroupHelperGen(OCIResourceHelperBase):
 
     def update_resource(self):
         update_details = self.get_update_model()
-        return oci_common_utils.call_with_backoff(
-            self.client.update_network_security_group,
-            network_security_group_id=self.module.params.get(
-                "network_security_group_id"
+        return oci_wait_utils.call_and_wait(
+            call_fn=self.client.update_network_security_group,
+            call_fn_args=(),
+            call_fn_kwargs=dict(
+                network_security_group_id=self.module.params.get(
+                    "network_security_group_id"
+                ),
+                update_network_security_group_details=update_details,
             ),
-            update_network_security_group_details=update_details,
+            waiter_type=oci_wait_utils.LIFECYCLE_STATE_WAITER_KEY,
+            operation=oci_common_utils.UPDATE_OPERATION_KEY,
+            waiter_client=self.client,
+            resource_helper=self,
+            wait_for_states=self.module.params.get("wait_until")
+            or oci_common_utils.get_resource_active_states(),
         )
 
     def delete_resource(self):
-        return oci_common_utils.call_with_backoff(
-            self.client.delete_network_security_group,
-            network_security_group_id=self.module.params.get(
-                "network_security_group_id"
+        return oci_wait_utils.call_and_wait(
+            call_fn=self.client.delete_network_security_group,
+            call_fn_args=(),
+            call_fn_kwargs=dict(
+                network_security_group_id=self.module.params.get(
+                    "network_security_group_id"
+                )
             ),
+            waiter_type=oci_wait_utils.LIFECYCLE_STATE_WAITER_KEY,
+            operation=oci_common_utils.DELETE_OPERATION_KEY,
+            waiter_client=self.client,
+            resource_helper=self,
+            wait_for_states=self.module.params.get("wait_until")
+            or oci_common_utils.get_resource_terminated_states(),
         )
 
 
@@ -290,6 +319,7 @@ def main():
         module=module,
         resource_type="network_security_group",
         service_client_class=VirtualNetworkClient,
+        namespace="core",
     )
 
     result = dict(changed=False)